BhoomiApp/backend/requirements.txt
absl-py==2.1.0
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.7
dataclasses-json==0.6.4
distro==1.9.0
docx==0.2.4
Flask==3.0.2
Flask-Cors==5.0.0
flatbuffers==24.3.25
frozenlist==1.4.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.66.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonify==0.5
jsonpatch==1.33
jsonpointer==2.4
keras==3.5.0
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-text-splitters==0.0.1
langsmith==0.1.45
libclang==18.1.1
lxml==5.2.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mathparse==0.1.2
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nltk==3.8.1
numpy==1.26.3
openai==0.28.0
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.0
packaging==23.2
pillow==10.4.0
Pint==0.23
protobuf==4.25.4
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.18.0
pymongo==3.13.0
python-dateutil==2.7.5
python-docx==1.1.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.8.0
scipy==1.12.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.29
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.4.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
yarl==1.9.4
